backend/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.15
faiss-cpu>=1.7.4
streamlit==1.29.0
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
openai>=1.6.1
python-dotenv==1.0.0
pydantic>=2.5.0
numpy>=1.21.0
